BACKEND/server-sacra360/OCR-service/requirements.txt
# OCR Service - Sacra360
# Dependencias específicas para el microservicio de OCR

# Framework web
fastapi==0.104.1
uvicorn==0.24.0

# Base de datos
sqlalchemy==1.4.46
psycopg2-binary==2.9.7

# OCR y procesamiento de imágenes
opencv-python==4.8.1.78
pytesseract==0.3.10
Pillow==10.0.1
numpy==1.24.3

# Validación y serialización
pydantic==1.10.13

# Utilidades
python-multipart==0.0.6  # Para manejar uploads
python-dotenv==1.0.0     # Para variables de entorno
requests==2.31.0         # Para testing

# Object Storage
minio==7.2.0            # Cliente para Minio object storage